reorganize lib
